requirements.txt
git+https://github.com/xr-coder/pyrofork
natsort
TgCrypto
pymongo
aiohttp
uvloop
python-dotenv
apscheduler
pytz